add more constraint

+++ weedcoco/repo/deposit.py
@@ -186,7 +186,7 @@ class RepositoryDataset:
                 rmtree(str(self.object_path))
 
     def remove(self):
-        if self.object_path.is_dir():
+        if self.object_path and self.object_path.is_dir():
             rmtree(str(self.object_path))
 
     def write_weedcoco(self, dataset_dir):
